Minor statistic changes

+++ SugarCenter/ViewModel/StatisticViewMovel.cs
@@ -3,7 +3,7 @@
     public class StatisticViewMovel
     {
         public int OrdersNumber { get; set; }
-        public int OrdersValue { get; set; }
+        public decimal OrdersValue { get; set; }
         public int BookingsNumber { get; set; }
     }
 }
